Default SDP bias or scaling value to be int16.
Int8 value isn't particularly useful for a per kernel bias or as a scaling factor. In most ML frameworks the bias is declared as int32 for low precision inteferencing so even int16 might be restrictive. However internally the SDP precision for int8/16 arthmetic is int32 so having a int32 bias value wouldn't work.